GestaoLivraria/pyproject.toml
[project]
name = "gestaolivraria"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicholas",email = "nicholasnas.nn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-conector-python (>=0.1.8,<0.2.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pytest (>=9.0.1,<10.0.0)"
]

[tool.poetry]
packages = [{include = "gestaolivraria", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-mock = "^3.15.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
